requirements.txt
mysql-connector-python
tornado>=4.2.1
PyYAML>=3.11
